+++ src/Extension.ml
@@ -2,6 +2,8 @@ open Import
 
 let selectSandboxCommandId = "ocaml.select-sandbox"
 
+let restartCommandId = "ocaml.server.restart"
+
 let openTerminalCommandId = "ocaml.open-terminal"
 
 let openTerminalSelectCommandId = "ocaml.open-terminal-select"
@@ -120,6 +122,19 @@ let selectSandbox (instance : Instance.t) () =
   in
   ()
 
+let restartInstance (instance : Instance.t) () =
+  let (_ : unit Promise.t) =
+    match instance.toolchain with
+    | None ->
+      selectSandbox instance ();
+      Promise.return ()
+    | Some toolchain ->
+      Instance.stop instance;
+      Instance.start instance toolchain
+      |> Promise.Result.iterError ~f:Window.showErrorMessage
+  in
+  ()
+
 let selectSandboxAndOpenTerminal () =
   let (_ : unit Promise.t) =
     Toolchain.select ()
@@ -128,6 +143,11 @@ let selectSandboxAndOpenTerminal () =
            Instance.openTerminal toolchain)
   in
   ()
+
+let openTerminal (instance : Instance.t) () =
+  match instance.toolchain with
+  | None -> selectSandboxAndOpenTerminal ()
+  | Some toolchain -> Instance.openTerminal toolchain
 
 let suggestToSetupToolchain instance =
   let open Promise.O in
@@ -144,16 +164,12 @@ let registerCommands extension =
       Vscode.ExtensionContext.subscribe extension
         (Vscode.Commands.register ~command ~handler))
 
-let openTerminal (instance : Instance.t) () =
-  match instance.toolchain with
-  | None -> selectSandboxAndOpenTerminal ()
-  | Some toolchain -> Instance.openTerminal toolchain
-
 let activate (extension : Vscode.ExtensionContext.t) =
   Js.Dict.set Process.env "OCAML_LSP_SERVER_LOG" "-";
   let instance = Instance.create () in
   registerCommands extension
     [ (selectSandboxCommandId, selectSandbox instance)
+    ; (restartCommandId, restartInstance instance)
     ; (openTerminalCommandId, openTerminal instance)
     ; (openTerminalSelectCommandId, selectSandboxAndOpenTerminal)
     ];