+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add highlighting for locally abstract types in OCaml files
 - Add highlighting for OASIS files
+- Improve OCamlbuild highlighting
 
 ## 0.7.0